refactor: restructure the project
- domain classes moved to common module
- renamed server module to kafka
- added song and user IDs to playback
- added idempotency to kafka producer
