+++ common/src/main/kotlin/pl/weronka/golonka/volatune/common/domain/Geohash.kt
@@ -1,0 +1,11 @@
+package pl.weronka.golonka.volatune.common.domain
+
+import ch.hsr.geohash.GeoHash
+
+typealias Geohash = String
+
+fun String.Companion.fromLatLng(
+    latitude: Double,
+    longitude: Double,
+    precision: Int = 7, // proximity of ~150m
+): Geohash = GeoHash.geoHashStringWithCharacterPrecision(latitude, longitude, precision)
