+++ common/src/main/kotlin/pl/weronka/golonka/volatune/common/domain/User.kt
@@ -1,0 +1,10 @@
+package pl.weronkagolonka.volatune.domain
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class User(
+    val id: String,
+    val name: String? = null,
+    val profileImageUrl: String? = null,
+)
