--- server/src/main/kotlin/pl/weronkagolonka/volatune/Main.kt
@@ -1,6 +1,0 @@
-package pl.weronkagolonka.volatune.pl.weronkagolonka.volatune
-
-fun main() {
-    val name = "DD"
-    println("Hello, $name!")
-}
