--- server/src/main/kotlin/pl/weronkagolonka/volatune/domain/Song.kt
@@ -1,11 +1,0 @@
-package pl.weronkagolonka.volatune.domain
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class Song(
-    val title: String,
-    val artists: List<String>,
-    val album: String,
-    val albumImageUrl: String,
-)
