--- server/src/main/kotlin/pl/weronkagolonka/volatune/domain/User.kt
@@ -1,9 +1,0 @@
-package pl.weronkagolonka.volatune.domain
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class User(
-    val name: String,
-    val profileImageUrl: String,
-)
